refactor(test_notebooks): move AttributeDictionary to tests
It's currently only used there.
